add security class to hash password, commented out for later use

+++ src/main/resources/schema.sql
@@ -5,7 +5,7 @@ DROP TABLE IF EXISTS container;
 DROP TABLE IF EXISTS user;
 
 CREATE TABLE user (
-	user_id INT NOT NULL AUTO_INCREMENT,
+	user_id BIGINT NOT NULL AUTO_INCREMENT,
 	user_name VARCHAR(128) NOT NULL,
 	user_password VARCHAR(128) NOT NULL,
 	user_email VARCHAR(128) NOT NULL,
@@ -14,29 +14,29 @@ CREATE TABLE user (
 );
 
 CREATE TABLE container (
-	container_id INT NOT NULL AUTO_INCREMENT,
+	container_id BIGINT NOT NULL AUTO_INCREMENT,
 	container_name VARCHAR(128) NOT NULL,
 	container_is_public boolean NOT NULL,
 	container_created_at VARCHAR(128),
-	user_id INT,
+	user_id BIGINT,
 	PRIMARY KEY (container_id),
 	FOREIGN KEY (user_id) REFERENCES user (user_id) ON DELETE CASCADE
 );
 
 
 CREATE TABLE topic (
-	topic_id INT NOT NULL AUTO_INCREMENT,
+	topic_id BIGINT NOT NULL AUTO_INCREMENT,
 	topic_value VARCHAR(128) NOT NULL,
 	topic_date VARCHAR(60) NOT NULL,
 	topic_created_by VARCHAR(60) NOT NULL,
 	topic_created_at VARCHAR(60),
-	container_id INT,
+	container_id BIGINT,
 	PRIMARY KEY (topic_id),
 	FOREIGN KEY (container_id) REFERENCES container (container_id) ON DELETE CASCADE
 );
 
 CREATE TABLE subtopic (
-	subtopic_id INT NOT NULL AUTO_INCREMENT,
+	subtopic_id BIGINT NOT NULL AUTO_INCREMENT,
 	subtopic_value VARCHAR(128) NOT NULL,
 	subtopic_date VARCHAR(60) NOT NULL,
 	subtopic_created_by VARCHAR(60) NOT NULL,
@@ -45,8 +45,8 @@ CREATE TABLE subtopic (
 );
 
 CREATE TABLE topic_subtopic (
-	topic_id INT NOT NULL,
-	subtopic_id INT NOT NULL,
+	topic_id BIGINT NOT NULL,
+	subtopic_id BIGINT NOT NULL,
 	FOREIGN KEY (topic_id) REFERENCES topic (topic_id) ON DELETE CASCADE,
 	FOREIGN KEY (subtopic_id) REFERENCES subtopic (subtopic_id) ON DELETE CASCADE
 );
